requirements.txt
Flask
pandas-gbq
plotly
gunicorn